Feature/HRP-6303 Change Estimated Tax field to decimal

* Modified EstimatedTax to decimal

* Added migrations

* Update client


Approved-by: Andrea Centeno
Approved-by: Wladimir

+++ Husa.Quicklister.Abor.Api.Contracts/Response/LotListing/LotFinancialResponse.cs
@@ -24,7 +24,7 @@ namespace Husa.Quicklister.Abor.Api.Contracts.Response.LotListing
         public DateTime? BonusExpirationDate { get; set; }
         public string HoaName { get; set; }
         public decimal? HoaFee { get; set; }
-        public int? EstimatedTax { get; set; }
+        public decimal? EstimatedTax { get; set; }
         public int? TaxYear { get; set; }
         public ICollection<TaxExemptions> TaxExemptions { get; set; }
         public int? TaxAssesedValue { get; set; }
