+++ Husa.Quicklister.Abor.Domain/Interfaces/LotListing/IProvideLotFinancial.cs
@@ -5,7 +5,7 @@ namespace Husa.Quicklister.Abor.Domain.Interfaces.LotListing
 
     public interface IProvideLotFinancial : IProvideCommonFinancial
     {
-        int? EstimatedTax { get; set; }
+        decimal? EstimatedTax { get; set; }
         int? TaxYear { get; set; }
         ICollection<TaxExemptions> TaxExemptions { get; set; }
         int? TaxAssesedValue { get; set; }
